requirements.txt
discord.py<1.7
python-decouple<3
